requirements.txt
streamlit
rapidocr_onnxruntime
PyMuPDF
numpy
Pillow
